feat: リサイズ処理を共通関数に集約
