+++ src/entry/avatar.js
@@ -7,8 +7,7 @@ import * as THREE from "three";
 import { CONFIG } from "../core/config.js";
 import { createControls } from "../core/controls.js";
 import { createPostPipeline } from "../core/postprocess.js";
-import { createRenderer,
-  fitToCanvas } from "../core/renderer.js";
+import { createRenderer, setupResize } from "../core/renderer.js";
 import { createSceneGraph } from "../core/scene.js";
 import { applyPS1Jitter, runFixedStepLoop } from "../core/utils.js";
 import { initBootOverlay } from "../core/boot-overlay.js";
@@ -22,18 +21,8 @@ const post = createPostPipeline(THREE, renderer, CONFIG); // { render(scene,came
 // ブートオーバーレイ初期化
 initBootOverlay();
 
-/**
- * Canvas サイズに合わせてレンダラー・ポスト処理・カメラを調整。
- */
-function resize()
-{
-  fitToCanvas(renderer, canvas, CONFIG);
-  post.resize(renderer);
-  camera.aspect = canvas.clientWidth / canvas.clientHeight;
-  camera.updateProjectionMatrix();
-}
-window.addEventListener("resize", resize);
-resize();
+// リサイズ処理の設定
+setupResize(renderer, canvas, camera, CONFIG, post);
 
 // ループ（PS1_MODE に合わせて分岐）
 const ROT_SPEED = 0.007 * 60;
